Move the AppStateManager to the data module
